Moved CounitngLongPredicate
